web: move status/rejection/New-game into a compact header bar
Extract the status, rejection, and New-game controls out of BoardView into a
StatusBarView rendered inline with the "Eldritch" heading (a flex .app-header
row), so they take one horizontal line instead of three stacked rows. The two
affected tests move to tests/status_bar.rs (mounting StatusBarView).
